+++ src/main/java/com/dong/service/LinkApplication.java
@@ -1,5 +1,7 @@
 package com.dong.service;
 
+import java.math.BigInteger;
+
 import org.dayatang.utils.Page;
 
 import com.dong.application.dto.LinkDTO;
@@ -8,6 +10,6 @@ public interface LinkApplication extends BaseApplication<LinkDTO, Long> {
 	
 	Page<LinkDTO> getPage(int currentPage, int pageSize);
 	
-	int getTotal();
+	BigInteger getTotal();
 
 }
